[IMP] purchase_stock_picking_show_currency_rate: reflect oca changes
