+++ purchase_stock_picking_show_currency_rate/tests/__init__.py
@@ -1,3 +1,4 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl)
 
+from . import common
 from . import test_purchase_stock_picking
